numb3rs/test_numb3rs.py
from numb3rs import validate

def main():
    test_validate()

def test_validate():
    assert validate("127.0.0.1") == True
    assert validate("134.134.134.134") == True
    assert validate("192.168.1.250") == True
    assert validate("512.512.512.512") == False
    assert validate("212.168.3.1") == True
    assert validate("0.0.0.0") == True
    assert validate("512.512.512.512.512") == False
    assert validate("234.193.45.263") == False
    assert validate("cat") == False
    assert validate("5.") == False
    assert validate("5") == False
    assert validate("5.45") == False
    assert validate("5.45.210") == False
    assert validate("275.45.67.4") == False
    assert validate("") == False
    assert validate(".") == False
    assert validate(" ") == False

if __name__ == "__main__":
    main()